+++ src/tools/_utils.ts
@@ -1,0 +1,9 @@
+import { PersistentHook, Wrapper } from "module-helpers";
+
+function toggleWrappersAndHooks(entries: (Wrapper | PersistentHook)[], enabled: boolean) {
+    for (const entry of entries) {
+        entry.toggle(enabled);
+    }
+}
+
+export { toggleWrappersAndHooks };